+++ lib/src/languages/java/java11/java11_grammar.dart
@@ -33,6 +33,8 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
         return ASTTypeArray(ASTTypeDynamic.instance);
       case 'var':
         return ASTTypeVar();
+      case 'final':
+        return ASTTypeVar(unmodifiable: true);
       default:
         return ASTType(name);
     }